Changed code to pull unique data from the Object_Path column instead of the ComputerName column
If Hyena has multiple runs against the same systems and concatenates the results it was shown that sometimes the ComputerName column would sometimes include a FQDN and other times not.  Because of this the -Unique option would treat them as two different systems.  Additionally, there were some servers that had an entry in the Object_Path but not in the ComputerName column.  This would result in those systems not be queried at all with USBDeview when taking unique ComputerName values.